+++ tests/algorithms/minimalistic/sample_main_minimalistic.py
